+++ lib/helper/web_oauth.dart
@@ -5,6 +5,7 @@ library msauth;
 
 import 'dart:async';
 import 'dart:convert';
+
 import 'package:aad_oauth/helper/core_oauth.dart';
 import 'package:aad_oauth/model/config.dart';
 import 'package:aad_oauth/model/failure.dart';
@@ -37,8 +38,8 @@ external Object jsGetAccessToken();
 @JS('getIdToken')
 external Object jsGetIdToken();
 
-@JS('isLogged')
-external bool jsIsLogged();
+@JS('hasCachedAccountInformation')
+external bool jsHasCachedAccountInformation();
 
 class WebOAuth extends CoreOAuth {
   final Config config;
@@ -65,9 +66,9 @@ class WebOAuth extends CoreOAuth {
         codeVerifier: config.codeVerifier,
         authorizationUrl: config.authorizationUrl,
         tokenUrl: config.tokenUrl,
+        cacheLocation: config.cacheLocation.value,
         customParameters: jsonEncode(config.customParameters),
-        postLogoutRedirectUri: config.postLogoutRedirectUri
-    ));
+        postLogoutRedirectUri: config.postLogoutRedirectUri));
   }
 
   @override
@@ -81,7 +82,8 @@ class WebOAuth extends CoreOAuth {
   }
 
   @override
-  Future<bool> get isLogged => Future<bool>.value(jsIsLogged());
+  Future<bool> get hasCachedAccountInformation =>
+      Future<bool>.value(jsHasCachedAccountInformation());
 
   @override
   Future<Either<Failure, Token>> login(