+++ lib/request_code.dart
@@ -36,6 +36,15 @@ class RequestCode {
 
     final webView = WebViewWidget(controller: controller);
 
+    if (_config.navigatorKey.currentState == null) {
+      throw Exception(
+        'Could not push new route using provided navigatorKey, Because '
+        'NavigatorState returned from provided navigatorKey is null. Please Make sure '
+        'provided navigatorKey is passed to WidgetApp. This can also happen if at the time of this method call '
+        'WidgetApp is not part of the flutter widget tree',
+      );
+    }
+
     await _config.navigatorKey.currentState!.push(
       MaterialPageRoute(
         builder: (context) => Scaffold(